requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
PyJWT==1.6.4